Break down capabilities mod into sub mods
